Handle basic.cancel frames on the connection
Previously it was not possible to cancel a consumer with basic_cancel,
and attempting to do so would hang the mock server.

Now basic.cancel messages are handled and responded to with
basic.cancelok.  The AmqpConnection._consumers attribute is now a dict
mapping (channel_id, consumer_tag) to each consumer task so they can
be looked up and cancelled.