allow binding a queue to an exchange
Previously, publishing a message to an exchange (either through the mock
client, or through a "real" client connected to the mock server) would
not result in that message being added to any queues bound to the
exchange.  Testing queues and exchanges were totally separate.  E.g. for
the same effect you would have to have a "real" client publish to an
exchange, and have the mock client publish to an associated queue.

This automatically adds any messages published to an exchange to any
queues bound to that exchange (this assumes direct routing--fanout
will be added in a follow-up).