support fanout exchange (#33)
* allow binding a queue to an exchange

Previously, publishing a message to an exchange (either through the mock
client, or through a "real" client connected to the mock server) would
not result in that message being added to any queues bound to the
exchange.  Testing queues and exchanges were totally separate.  E.g. for
the same effect you would have to have a "real" client publish to an
exchange, and have the mock client publish to an associated queue.

This automatically adds any messages published to an exchange to any
queues bound to that exchange (this assumes direct routing--fanout
will be added in a follow-up).

* two more enhancements concerning queues and exchanges:

declare_queue now actually does something, in terms of ensuring that the
queue exists

new queues are automatically bound to the default exchange

* add support for fanout exchanges

topic/header exchanges can be left as a separate exercise

depends on https://github.com/nikitanovosibirsk/amqp-mock/pull/8
and https://github.com/nikitanovosibirsk/amqp-mock/pull/9

* fix linter

* add tests

* Change storage history from dict to list to support multiple copies of the same message with fanout exchanges

* Change test_routing_fanout_exchange to test that bound queues receive message regardless of routing key

* Remove python3.7 support

---------

Co-authored-by: E. Madison Bray <erik.bray@lri.fr>
Co-authored-by: Nikita Tsvetkov <nikitanovosibirsk@yandex.com>